更新注释，优化代码结构
