+++ python/__init__.py
@@ -173,7 +173,7 @@ from appbuilder.core.agent import AgentRuntime
 from appbuilder.core.user_session import UserSession
 
 from appbuilder.utils.logger_util import logger
-from appbuilder.core.manifest.models import function_to_manifest,decorator_to_manifest
+from appbuilder.core.manifest.manifest import function_to_manifest,decorator_to_manifest
 from appbuilder.core.manifest.manifest_decorator import ManifestView, manifest, manifest_parameter, manifest_return
 
 from appbuilder.core.utils import get_model_list
