+++ python/core/manifest/manifest_decorator.py
@@ -11,8 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""ManifestView classes."""
-
 import inspect
 import re
 from enum import Enum
