+++ python/core/manifest/manifest_signature.py
@@ -1,6 +1,16 @@
-# -*- coding: UTF-8 -*-
-"""Parse function parameters from signature."""
-
+# Copyright (c) 2024 Baidu, Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import inspect
 from inspect import Parameter, Signature
 from typing import Any, Dict, Union, Optional, List, get_origin, get_args
